+++ d2/src/main/java/com/linkedin/d2/balancer/strategies/relative/StateUpdater.java
@@ -313,9 +313,17 @@ public class StateUpdater
       }
       else
       {
-        // If it is a new client, we directly set health score as the initial health score to initialize
-        trackerClientStateMap.put(trackerClient, new TrackerClientState(_relativeStrategyProperties.getInitialHealthScore(),
-            _relativeStrategyProperties.getMinCallCount()));
+        // Initializing a new client score
+        if (trackerClient.doNotSlowStart())
+        {
+          trackerClientStateMap.put(trackerClient, new TrackerClientState(MAX_HEALTH_SCORE,
+              _relativeStrategyProperties.getMinCallCount()));
+        }
+        else
+        {
+          trackerClientStateMap.put(trackerClient,
+              new TrackerClientState(_relativeStrategyProperties.getInitialHealthScore(), _relativeStrategyProperties.getMinCallCount()));
+        }
       }
     }
     partitionState.setPartitionStats(avgClusterLatency, clusterCallCount, clusterErrorCount);
@@ -433,6 +441,7 @@ public class StateUpdater
     if (LOG.isDebugEnabled())
     {
       LOG.debug("Strategy updated: service=" + _serviceName
+          + ", partitionId=" + partitionId
           + ", unhealthyClientNumber=" + newUnhealthyClients.size()
           + ", newState=" + newState
           + ", unhealthyClients={" + (newUnhealthyClients.stream().limit(LOG_UNHEALTHY_CLIENT_NUMBERS)
@@ -443,7 +452,8 @@ public class StateUpdater
     }
     else if (allowToLog(oldState, newState, newUnhealthyClients, oldUnhealthyClients))
     {
-      LOG.debug("Strategy updated: service=" + _serviceName
+      LOG.info("Strategy updated: service=" + _serviceName
+          + ", partitionId=" + partitionId
           + ", unhealthyClientNumber=" + newUnhealthyClients.size()
           + ", newState=" + newState
           + ", unhealthyClients={" + (newUnhealthyClients.stream().limit(LOG_UNHEALTHY_CLIENT_NUMBERS)