+++ data/src/main/java/com/linkedin/data/DataList.java
@@ -17,7 +17,6 @@
 package com.linkedin.data;
 
 import com.linkedin.data.collections.CheckedList;
-import com.linkedin.data.collections.CommonList;
 import com.linkedin.data.collections.ListChecker;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -97,7 +96,7 @@ public final class DataList extends CheckedList<Object> implements DataComplex
     o._instrumented = false;
     o._accessList = null;
     o._dataComplexHashCode = 0;
-    o._isTraversing = new ThreadLocal<>();
+    o._isTraversing = null;
 
     return o;
   }
@@ -273,13 +272,31 @@ public final class DataList extends CheckedList<Object> implements DataComplex
 
   private final static ListChecker<Object> _checker = (list, e) -> Data.checkAllowed((DataComplex) list, e);
 
+  ThreadLocal<Object> isTraversing()
+  {
+    if (_isTraversing == null)
+    {
+      synchronized (this)
+      {
+        if (_isTraversing == null)
+        {
+          _isTraversing = new ThreadLocal<>();
+        }
+      }
+    }
+
+    return _isTraversing;
+  }
+
   /**
    * Indicates if this {@link DataList} is currently being traversed by a {@link Data.TraverseCallback} if this value is
    * not null, or not if this value is null. This is internally marked package private, used for cycle detection and
    * not meant for use by external callers. This is maintained as a {@link ThreadLocal} to allow for concurrent
    * traversals of the same {@link DataList} from multiple threads.
+   *
+   * <p>This variable is lazy instantiated since ThreadLocal instantiation can be expensive under thread contention.</p>
    */
-  ThreadLocal<Object> _isTraversing = new ThreadLocal<>();
+  private volatile ThreadLocal<Object> _isTraversing = null;
 
   private boolean _madeReadOnly = false;
   private boolean _instrumented = false;