+++ gradle-plugins/src/main/java/com/linkedin/pegasus/gradle/tasks/GenerateDataTemplateTask.java
@@ -17,6 +17,7 @@ import org.gradle.api.tasks.CacheableTask;
 import org.gradle.api.tasks.Classpath;
 import org.gradle.api.tasks.Input;
 import org.gradle.api.tasks.InputDirectory;
+import org.gradle.api.tasks.Optional;
 import org.gradle.api.tasks.OutputDirectory;
 import org.gradle.api.tasks.PathSensitive;
 import org.gradle.api.tasks.PathSensitivity;
@@ -40,11 +41,15 @@ import static com.linkedin.pegasus.gradle.SharedFileUtils.getSuffixedFiles;
 @CacheableTask
 public class GenerateDataTemplateTask extends DefaultTask
 {
-  private File _destinationDir;
+  // Input Task Property
   private File _inputDir;
   private FileCollection _resolverPath;
   private FileCollection _codegenClasspath;
   private boolean _enableArgFile;
+  private Boolean _generateLowercasePath;
+
+  // Output Task Property
+  private File _destinationDir;
 
   public GenerateDataTemplateTask()
   {
@@ -121,6 +126,18 @@ public class GenerateDataTemplateTask extends DefaultTask
     _enableArgFile = enable;
   }
 
+  @Optional
+  @Input
+  public Boolean generateLowercasePath()
+  {
+    return _generateLowercasePath;
+  }
+
+  public void setGenerateLowercasePath(Boolean enable)
+  {
+    _generateLowercasePath = enable;
+  }
+
   @TaskAction
   public void generate()
   {
@@ -167,6 +184,10 @@ public class GenerateDataTemplateTask extends DefaultTask
       javaExecSpec.setMain("com.linkedin.pegasus.generator.PegasusDataTemplateGenerator");
       javaExecSpec.setClasspath(_pathedCodegenClasspath);
       javaExecSpec.jvmArgs("-Dgenerator.resolver.path=" + resolverPathArg);
+      if (_generateLowercasePath != null)
+      {
+        javaExecSpec.jvmArgs("-Dgenerator.generate.lowercase.path=" + _generateLowercasePath); //.run(generateLowercasePath)
+      }
       javaExecSpec.jvmArgs("-Droot.path=" + getProject().getRootDir().getPath());
       javaExecSpec.args(_destinationDir.getPath());
       javaExecSpec.args(_inputDir);